+++ server/public/error.html
@@ -4,7 +4,7 @@
         <meta charset="UTF-8">
         <meta name="viewport" content="width=device-width, initial-scale=1.0">
         <meta http-equiv="X-UA-Compatible" content="ie=edge">
-        <title>Blindtest - 404</title>
+        <title>XD2A3 - Error</title>
         <link rel="stylesheet" href="/public/style.css">
     </head>
     <body>
